Add ShipSalesOrder endpoint and support fractional quantities
A new ShipSalesOrder endpoint was added to the SalesController
to handle the shipment of sales orders. This includes stock
management, fulfillment updates, and transaction logging.

The TotalQuantity and TotalFulfilledQuantity properties in the
SalesOrder class were updated from int to decimal to support
fractional quantities. A migration was added to update the
SalesOrders table schema accordingly.

The ManufacturingDbContextModelSnapshot was updated to reflect
the schema changes. A new helper method, UpdateSalesOrderItemsStatus,
was added to manage item-level status updates based on fulfillment
progress.

+++ ERP/Entity/Document/SalesOrder.cs
@@ -20,8 +20,8 @@ public class SalesOrder
     public string? CustomerEmail { get; set; }
 
     //  NEW: Track fulfillment summary
-    public int TotalQuantity { get; set; }
-    public int TotalFulfilledQuantity { get; set; }
+    public decimal TotalQuantity { get; set; }
+    public decimal TotalFulfilledQuantity { get; set; }
 
     //  NEW: Optional link to Purchase Order (if 1 PO per SO)
     public Guid? PurchaseOrderId { get; set; }
